102303945-Web-Service/requirements.txt
Flask
pandas
python-dotenv
numpy
gunicorn
